+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project are documented here. The format is based on
 
 ## [Unreleased]
 
+## [0.5.0] - 2026-07-11
+
 ### Added
 - shunt now creates bot-only branch protection for its `mq/<base>/staging*`
   staging refs.
@@ -132,7 +134,8 @@ Initial release.
   non-interactive Git credential prompts instead of embedding tokens in clone
   URLs.
 
-[Unreleased]: https://github.com/rbtr/shunt/compare/v0.4.1...HEAD
+[Unreleased]: https://github.com/rbtr/shunt/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/rbtr/shunt/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/rbtr/shunt/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/rbtr/shunt/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/rbtr/shunt/compare/v0.2.0...v0.3.0